Phuong | Move the Earth to different file
